Rename: dto.user.auth로 auth 관련 dto 이동
