+++ src/main/java/com/teamddd/duckmap/controller/AuthController.java
@@ -8,9 +8,9 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.teamddd.duckmap.dto.LoginUserReq;
-import com.teamddd.duckmap.dto.LoginUserRes;
 import com.teamddd.duckmap.dto.Result;
+import com.teamddd.duckmap.dto.user.auth.LoginUserReq;
+import com.teamddd.duckmap.dto.user.auth.LoginUserRes;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.parameters.RequestBody;
