+++ src/main/java/com/teamddd/duckmap/dto/user/auth/LoginUserReq.java
@@ -1,4 +1,4 @@
-package com.teamddd.duckmap.dto;
+package com.teamddd.duckmap.dto.user.auth;
 
 import javax.validation.constraints.NotBlank;
