+++ src/main/java/com/teamddd/duckmap/dto/user/auth/LoginUserRes.java
@@ -1,4 +1,4 @@
-package com.teamddd.duckmap.dto;
+package com.teamddd.duckmap.dto.user.auth;
 
 import lombok.Builder;
 import lombok.Getter;
